Smithery/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
mcp-server==1.0.0
requests==2.31.0
python-dotenv==1.0.0
fastmcp==0.1.0